+++ model/relation/event/processor/EventProcessorInterface.php
@@ -23,24 +23,11 @@ declare(strict_types=1);
 namespace oat\taoMediaManager\model\relation\event\processor;
 
 use oat\oatbox\event\Event;
-use oat\oatbox\service\ConfigurableService;
-use oat\taoMediaManager\model\relation\event\MediaRemovedEvent;
-use oat\taoMediaManager\model\relation\service\ItemRelationUpdateService;
 
-class MediaRemovedProcessor extends ConfigurableService implements ProcessorInterface
+interface EventProcessorInterface
 {
-    public function process(Event $event): void
-    {
-        if (!$event instanceof MediaRemovedEvent) {
-            throw new InvalidEventException($event);
-        }
-
-        $this->getItemRelationUpdateService()
-            ->removeMedia($event->getMediaId());
-    }
-
-    private function getItemRelationUpdateService(): ItemRelationUpdateService
-    {
-        return $this->getServiceLocator()->get(ItemRelationUpdateService::class);
-    }
+    /**
+     * @throws InvalidEventException
+     */
+    public function process(Event $event): void;
 }
