+++ test/unit/model/relation/event/ItemUpdatedProcessorTest.php
@@ -26,13 +26,13 @@ use oat\generis\test\TestCase;
 use oat\oatbox\event\Event;
 use oat\taoItems\model\event\ItemUpdatedEvent;
 use oat\taoMediaManager\model\relation\event\processor\InvalidEventException;
-use oat\taoMediaManager\model\relation\event\processor\ItemUpdatedProcessor;
+use oat\taoMediaManager\model\relation\event\processor\ItemUpdatedEventProcessor;
 use oat\taoMediaManager\model\relation\service\ItemRelationUpdateService;
 use PHPUnit\Framework\MockObject\MockObject;
 
 class ItemUpdatedProcessorTest extends TestCase
 {
-    /** @var ItemUpdatedProcessor */
+    /** @var ItemUpdatedEventProcessor */
     private $subject;
 
     /** @var ItemRelationUpdateService|MockObject */
@@ -41,7 +41,7 @@ class ItemUpdatedProcessorTest extends TestCase
     public function setUp(): void
     {
         $this->updateService = $this->createMock(ItemRelationUpdateService::class);
-        $this->subject = new ItemUpdatedProcessor();
+        $this->subject = new ItemUpdatedEventProcessor();
         $this->subject->setServiceLocator(
             $this->getServiceLocatorMock(
                 [
